requirements.txt
ortools
pyjobshop
python-dotenv
mysql-connector-python
flask
PyQt5